Pushed the fragment drawer on accept down.
